refactor: changed projects page structure
