+++ pages/projects/[projectId].tsx
@@ -1,0 +1,5 @@
+const ProjectDetails = () => {
+  return <></>;
+};
+
+export default ProjectDetails;
